Added RDoc documentation to new files.

+++ R/blast.R
@@ -1,9 +1,14 @@
-parse_blast_table <- function(filepath) {
+#' Parse tabular output from BLAST
+#'
+#' @param filepath Path to tabular BLAST output file.
+#' @return A data frame of BLAST results.
+#' @export
+parse_blast_table <- function (filepath) {
   column_classes <- c(
     "character", "character", "numeric",
     "integer", "integer", "integer",
     "integer", "integer", "integer", "integer",
-    "numeric", "integer"
+    "numeric", "numeric"
   )
   column_names <- c(
     "query_id", "subject_id", "pct_identity",
